[DEV-3384] refactor 3
