+++ src/components/payment/add-bank-account.tsx
@@ -1,9 +1,10 @@
-import { BankAccount, CreateBankAccount, Utils, Validations, useBankAccountContext } from '@dfx.swiss/react';
+import { ApiError, BankAccount, CreateBankAccount, Utils, Validations, useBankAccountContext } from '@dfx.swiss/react';
 import {
   Form,
   StyledButton,
   StyledButtonColor,
   StyledButtonWidth,
+  StyledInfoText,
   StyledInput,
   StyledSpacer,
   StyledVerticalStack,
@@ -15,12 +16,15 @@ import { ErrorHint } from '../error-hint';
 
 interface AddBankAccountProps {
   onSubmit: (bankAccount: BankAccount) => void;
+  confirmationText?: string;
 }
 
-export function AddBankAccount({ onSubmit }: AddBankAccountProps): JSX.Element {
+export function AddBankAccount({ onSubmit, confirmationText }: AddBankAccountProps): JSX.Element {
   const { translate, translateError } = useSettingsContext();
 
   const [error, setError] = useState<string>();
+  const [customError, setCustomError] = useState<string>();
+  const [confirmBankAccount, setConfirmBankAccount] = useState<BankAccount>();
 
   const {
     control,
@@ -32,16 +36,52 @@ export function AddBankAccount({ onSubmit }: AddBankAccountProps): JSX.Element {
 
   async function createBankAccount(newAccount: CreateBankAccount): Promise<void> {
     setError(undefined);
+    setCustomError(undefined);
+    setConfirmBankAccount(undefined);
+
     createAccount(newAccount)
-      .then(onSubmit)
-      .catch((e) => setError(e.message ?? 'Unknown error'));
+      .then(!confirmationText ? onSubmit : setConfirmBankAccount)
+      .catch((e: ApiError) => {
+        if (e.statusCode === 403) {
+          setCustomError(translate('screens/iban', 'This IBAN already exists in another DFX customer account.'));
+        } else if (e.statusCode === 409) {
+          let error = translate('screens/iban', 'This IBAN already exists in another DFX customer account.') + ' ';
+          error += e.message?.includes('account merge')
+            ? translate(
+                'screens/kyc',
+                'We have just sent you an email. To continue with your existing account, please confirm your email address by clicking on the link sent.',
+              )
+            : translate('screens/kyc', 'Start the KYC process with the same email to merge your accounts.');
+
+          setCustomError(error);
+        } else if (e.statusCode === 400 && e.message?.includes('Multi-account IBAN')) {
+          setCustomError(
+            translate(
+              'screens/iban',
+              'This is a multi-account IBAN and cannot be added as a personal account. Please send the confirmation of the bank transaction as a PDF to support@dfx.swiss.',
+            ),
+          );
+        } else {
+          setError(e.message ?? 'Unknown error');
+        }
+      });
   }
 
   const rules = Utils.createRules({
     iban: [Validations.Required, Validations.Iban(allowedCountries)],
   });
 
-  return (
+  return confirmBankAccount ? (
+    <>
+      <p className="text-dfxGray-700">{confirmationText}</p>
+
+      <StyledButton
+        color={StyledButtonColor.RED}
+        label={translate('general/actions', 'OK')}
+        onClick={() => onSubmit(confirmBankAccount)}
+      />
+    </>
+  ) : (
     <Form
       control={control}
       rules={rules}
@@ -50,7 +90,11 @@ export function AddBankAccount({ onSubmit }: AddBankAccountProps): JSX.Element {
       translate={translateError}
     >
       <StyledVerticalStack gap={4}>
-        <StyledInput name="iban" label={translate('screens/payment', 'IBAN')} placeholder="XX XXXX XXXX XXXX XXXX X" />
+        <StyledInput
+          name="iban"
+          label={translate('screens/payment', 'IBAN')}
+          placeholder={translate('screens/payment', 'CH46 8914 4632 3427 5387 5')}
+        />
         <StyledInput
           name="label"
           label={translate('screens/sell', 'Optional - Account Designation')}
@@ -61,6 +105,12 @@ export function AddBankAccount({ onSubmit }: AddBankAccountProps): JSX.Element {
         {error && (
           <div className="text-center">
             <ErrorHint message={error} />
+          </div>
+        )}
+
+        {customError && (
+          <div className="text-left">
+            <StyledInfoText invertedIcon>{customError}</StyledInfoText>
           </div>
         )}
